Refactored MockDigitalInput and moved PushButtonHelperTests to the correct package
