+++ pi4micronaut-utils/src/test/java/com/opensourcewithslu/mock/MockDigitalInput.java
@@ -1,4 +1,4 @@
-package com.opensourcewithslu.inputdevices;
+package com.opensourcewithslu.mock;
 import com.pi4j.common.Metadata;
 import com.pi4j.context.Context;
 import com.pi4j.exception.InitializeException;
@@ -6,6 +6,13 @@ import com.pi4j.exception.ShutdownException;
 import com.pi4j.io.binding.DigitalBinding;
 import com.pi4j.io.gpio.digital.*;
 
+/*
+The ockDigitalInput class implements the DigitalInput interface, which requires
+all methods to be implemented, even if they are not used in the mock.
+These methods have default or no operation implementations to fulfill the interface contract.
+they provided an empty method("return null") because the methods are required by the interface,
+yet for the mock implementation, it doesn't need to do anything.
+*/
 public class MockDigitalInput implements DigitalInput {
     DigitalState mockState;
     public MockDigitalInput(){
@@ -71,7 +78,7 @@ public class MockDigitalInput implements DigitalInput {
 
     @Override
     public String name() {
-        return "";
+        return "MockDigitalInput";
     }
 
     @Override
